Reuse GetMethod from handler, properly synchronize calls to timeout handler, and support mapping
