+++ src/NServiceBus.Testing/Handler.cs
@@ -2,9 +2,6 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Linq.Expressions;
-    using System.Threading.Tasks;
     using NServiceBus.MessageInterfaces.MessageMapper.Reflection;
     using NServiceBus.Testing.ExpectedInvocations;
 
@@ -276,7 +273,7 @@
             testableMessageHandlerContext.MessageId = messageId;
 
             var messageType = messageCreator.GetMappedTypeFor(message.GetType());
-            var handleMethod = GetMethod(handler.GetType(), messageType, typeof(IHandleMessages<>));
+            var handleMethod = handler.GetType().CreateInvoker(messageType, typeof(IHandleMessages<>));
             if (handleMethod == null)
             {
                 return;
@@ -331,36 +328,9 @@
             throw new NotImplementedException();
         }
 
-        static Func<object, object, IMessageHandlerContext, Task> GetMethod(Type targetType, Type messageType, Type interfaceGenericType)
-        {
-            var interfaceType = interfaceGenericType.MakeGenericType(messageType);
 
-            if (!interfaceType.IsAssignableFrom(targetType))
-            {
-                return null;
-            }
 
-            var methodInfo = targetType.GetInterfaceMap(interfaceType).TargetMethods.FirstOrDefault();
-            if (methodInfo == null)
-            {
-                return null;
-            }
-
-            var target = Expression.Parameter(typeof(object));
-            var messageParam = Expression.Parameter(typeof(object));
-            var contextParam = Expression.Parameter(typeof(IMessageHandlerContext));
-
-            var castTarget = Expression.Convert(target, targetType);
-
-            var methodParameters = methodInfo.GetParameters();
-            var messageCastParam = Expression.Convert(messageParam, methodParameters.ElementAt(0).ParameterType);
-
-            Expression body = Expression.Call(castTarget, methodInfo, messageCastParam, contextParam);
-
-            return Expression.Lambda<Func<object, object, IMessageHandlerContext, Task>>(body, target, messageParam, contextParam).Compile();
-        }
-
-        private readonly T handler;
+        T handler;
 
         MessageMapper messageCreator = new MessageMapper();
